Refactor and optimize the full-stack application.
This commit introduces a comprehensive set of enhancements to both the smart contract and the frontend, focusing on performance, maintainability, and user experience.

**Smart Contract (FrameBreaker.sol):**
- Replaced `require` statements with custom errors for improved gas efficiency and clearer error handling.
- Made `MAX_LEADERBOARD_SIZE` an `immutable` constant to reduce deployment and runtime gas costs.
- Simplified the constructor by removing the redundant `_initialOwner` parameter.
- Updated tests to align with the new custom errors and constructor signature.

**Frontend:**
- Refactored the monolithic `App.tsx` into smaller, more manageable components (`GameUI.tsx`, `GameCanvas.tsx`) and a custom hook (`useGameLogic.ts`) to separate concerns.
- Created a server-side API endpoint for AI level generation to resolve a critical client-side rendering issue caused by the `@google/genai` library.
- Redesigned the "Game Over" screen with a more intuitive two-column layout for a better user experience.
- Fixed a bug in the game logic where the ball would not automatically relaunch after a life was lost.

**Build & Configuration:**
- Corrected the `foundry.toml` configuration to ensure tests are discovered and dependencies are correctly mapped.
- Updated `vite.config.ts` to support the new server-side API endpoint and correctly load environment variables.
